MVC - add exportCsv() to ApiControllerBase to easily stream recordsets as csv files to the client and add asRecordSet() to ArrayField to be able to use this easily.
Sometimes it's practical to be able to use tools like excel to modify data, in order to do this we need an import and an export option in the model.
The export seems to be quite easy, controllers and forms can easily use this with a construction like:

$this->exportCsv($mymodel->myarray->asRecordSet());

The import is likely a bit more challenging.