mvc: "make validate" was not validating the full model...
... which was the intention of this helper.  Now, we see a lot of
other things that need fixing, but there is an implication for
migrations ("make migrate") where the copy of that function came
from...  Apparenlty migration only validates the model changes done
during migration, which is fine, but if we are restricting validation
further during a migration and forget to change these fields they are
not marked for validation and thus "validate" "correctly" leaving them
in a broken state in the config.  This can be cross-checked by the
GUI attempting to save these faulty values which is then no longer
possible.  Needs more discussion.