dhcrelay: functional replacement #6983
Since we are not ISC we can make a new menu spot.  Keep the legacy
reconfiguration spot as we end up building a complex command line
to run and this doesn't have to go into a template that just stuffs
args into a variable to start it, which then requires and rc.d file.

Though we want to move it out of dhcpd.inc at some point.

Apparently the new MVC page is missing an apply button.  ;)