firmware: avoid calculating + 0 in previous
fetch(1) appears to return "Unknown" when not being able to
get the size of the file from the mirror.  It's a bit odd
considering the code in question didn't change for years on
all ends and so far nobody reported a similar issue.

The calculation isn't strictly necessary as we try to fill
the gap between pkg integration and extra sets that should
be handled by pkg-base eventually.