interfaces: force regeneration of link-local on spoofed MAC; closes #4430
While here make it possible to go back to the original MAC address by
providing the "hwaddr" field of ifconfig back to link command.

This operates under the constraint that the interface where the MAC
address changes is the same as the interface where IPv6 is running
so that we do not perform useless or problematic configuration.

The case for VLANs and parent MACs is probably also quite tricky, but
since it has never been considered before do not start now.