mvc: partially fix container field cloning, closes https://github.com/opnsense/core/issues/7521
While cloning model fields, the identifier for these new fields should also be set, which is currently being arranged in the ArrayField->add() method.
This commit makes sure we set the identifiers for one level below the array field type, which is good enough for current implementations, eventually this should be a recursive action or needs a way to push the identifiers directly on clone.

re-arrange the code a bit as well to remove the need for the $new_record array.