System: Trust: Certificates - export p12 resulted in mangled binary blob. closes https://github.com/opnsense/core/issues/7703
It looks like out atob() resulted in sending broken data back to the client, the easiest way to prevent this is to keep our endpoint b64 encoded data as is and let the client handle it.
Accoring to https://developer.mozilla.org/en-US/docs/Web/HTTP/Basics_of_HTTP/Data_URLs, we should be able to suffix ;base64 in these cases.

While here, also make sure the result status is as expected.