Reporting / rrd - start using cron for rrd collection.
This is a temporary solution, but a first step into cleaning up rrd stat collection.

In this first milestone, we keep using the generated script, but remove the loop and sleep construct out of it, offering the posibility to handle service control to cron. To prevent the script being written at the same time we're collecting, we always lock the script before use.

should fix: https://github.com/opnsense/core/issues/7753