interfaces: replace 'newwanip_map' from interface apply
This always bugged me: 'newwanip_map' is called when e.g. DHCP is
done but here we just do it async to start something.  This doesn't
make sense, because eventually it triggers twice.

What rc.reload_all and setports.php do is call 'local' and I think
that makes sense.