firmware: move bogons fetch to firmware scripts
Be done with the retry and use the launcher.sh -r argument for
the cron-based execution.  Leave the backend commands where they
are.

Since the file had no copyright header looking at the old code
it looks like Bill and Scott worked on the file first in 2005.
Not really concerned with consecutive history here for lack of
creating a copyright header for so long.