mvc: implement reusable grid template using form definitions. for https://github.com/opnsense/core/issues/8187
Although this might not be the final version, it adds the relevant metadata to the form definitions and offers a very simple template which can be used as a partial().

Example input field:

    <field>
        <id>sysctl.tunable</id>
        <label>Tunable</label>
        <type>text</type>
        <grid_view>
            <sequence>1</sequence>
        </grid_view>
    </field>

grid_view entries contain a sequence which may overwrite the default order and can overwrite any data-* attribute available, e.g. <sortable>false</sortable> would remove the sortable option from the field.