Firewall: Automation: Filter - add (experimental) shaper support in rules
pf in FreeBSD 14.x supports dnpipe and dnqueue targets which hook dummynet, in theory we should be able to use these as well although our solution uses ipfw
for the shaper.

There might be some risks involved when using shapers in pf and configuring rules in ipfw, which is also why this feature is marked experimental.

Benefits of offering shaper targets in our standard firewall are increased flexibility (use of large aliases for example) and cleaner rulesets for example.

A quick test of a rule to shape iperf3 traffic does seem to work as expected.