mvc:model - support lazy loading (skip_dynamic_info) on alias models and use it in NetworkAliasField to improve performance by ignoring fields we're not going to use anyway.
The configd call "filter diag table_size" is quite convenient when inspecting the aliases grid, but if we're only interested in the names and descriptions, it adds quite some overhead.
We can skip the collection of this information by constructing with $skip_dynamic_info set to "true" now, which increases performance quite a bit.