dnsmasq: migrate to MVC/API, use interface binding instead of trying to search addresses in legacy backend. https://github.com/opnsense/core/issues/7905
Introduced ages ago (https://github.com/pfsense/pfsense/commit/b4323f39bc7c942b4a98943057f45cca7c701109), it's time to say goodbye to this trickery.
When interfaces are specified, dnsmasq binds to all (*) and filters the requests it should answer, which makes the bind more resilient to change as well (and shouldn't require a restart on interface changes).

This change is intended to ease the migration with minimal impact on current setups.