Firewall: Aliases - performance improvements when using large amounts of aliases, for https://github.com/opnsense/core/issues/8654
Add a handle to flush cached data for a model, specifically in cases where a model also contains dynamic data, the responsible component for that data is also responsible for flushing its cache.
A concrete example for this is the captive portal hooking into the array class of our aliases, in which case the model data can not reliably know the state of the cache when looking at the stored data.