Services: Unbound DNS - small experiment to improve https://github.com/opnsense/core/issues/8439
When an excessive number of events is fired, unbound may keep reconfiguring and keeps mounting and unmounting the requirements for our python plugin.
In order to prevent this, we should consider keeping the mounts in place on restarts when we are going to remount them anyway.

This likely won't solve all issues in the world, but may solve some of them. Ideally unbound shouldn't restart on interface changes, but that either requires unbound-control using the reload command (which is experimental) or preventing the configuration to include dynamic parts.