OPNsense/Auth - add minimalistic interface for SSO support.
In order to support other authentication methods for the business edition in the future, we need some glue to offer room for a different authentication flow.
When using an SSO flow, the service binding is usually determined by the endpoint which acts as a trampoline.

This commits mildly refactors our AuthenticationFactory and adds some handles to be able to choose SSO providers when available (using our standard listServers() call)

ISSOContainer describes the minimal interface a container of providers should support, the Provider class is a "struct" type container offering relevant information for the provider in question.