tabulator: more stability fixes
This is a backport of https://github.com/opnsense/core/pull/8795/commits/5026dd438377456f71cc4ed31ea9255b7b57d5c6.
Includes the removal of the old jquery bootgrid includes.

changes on the tabulator side:
- ambiguous 'datakey' usage has been fixed, the original intent
was mixed up with 'data-identifier'. However, 'datakey' is only
meant for command action identification.
- As a continuation of the above, the overridding of action commands
has now been improved for usage in the virtualDOM model. Instead
of using a MutationObserver to watch for inserted elements, the
responsibility has been delegated to onRendered for all formatters,
including custom commands. As long as the rendered HTML contains
the bootgrid-tooltip class, the tooltips will be activated once rendered.

For the system activity, routes and voucher pages, no UIBootgrid was used,
as ajax was set to false in these cases. Since UIBootgrid requires a 'search'
property to initialize, this commit changs this behavior slightly to not require
this property if ajax=false. This passes to jquery's original bootgrid as well,
but makes sure that the pages use the new tabulator format.