mvc: slightly improve replaceInputWithSelector() to support an empty placeholder, which should help for https://github.com/opnsense/core/pull/9158
This change uses a token for the single/manual input, in which case callers may use definitions like:

......
            'networks' => [
                'label' => gettext("Networks"),
                'items' => [
                    '' => gettext('any'),
                    '(self)' => gettext("This Firewall")
                ]
            ]
.....

When empty is offered, "any" will be choosen in this case. In theory we could change the filter model as well as "any" has no functional meaning anyway, but for simplicty this only changes the javascript part.