backend: make mwexecf*() independent from mwexec[^f]* #9325
mexec(), mwexec_bg() and mwexecf_bg() are now deprecated and
should be removed.  mwexecf_bg() is a bit of an oddity here but
since we now have mwexecfn() it is better used as mwexecfb()
and the error muting was also disabled on this one instead of
disabling it on mwexecf_bg().

This allows us to introduce this code in a batch update and
cherry-pick the callers afterwards.