system: use is_int()/array_key_first() in toArray() and fromArray() #9485
The approximation of the magic here is that we are looking for array
elements created by a natural append [] = or equivalent which has
an integer key of a rough range of 0 to count() - 1, but not always as
we can see from the ticket.

unset() breaks the pledge of sequential lists and makes array_is_list()
fail.  Sorting would also break the sequential pledge without resetting
the keyes using array_values() but that approach is too broad.

Instead, get a single key we can do a strict int type check on so that
we are as likely to succeed as was the case before the change in 9106a6f.

It's also fast.  ;)