tests: unwind BaseModel test side effect on AclConfig
Give the test its own empty config to migrate into, which makes
resetting it irrelevant.  A tiny bit of magic is needed to bootstrap
a config.xml so it's not the one that is checked in and we also
remove the temporary files from the tree so it does not confuse git.

save() is a far-reaching side effect in the test, but more or less
necessary for better test coverage.  But I'd still want a simpler
test around Core\Config and this offers merely offers the groundwork
also removing the need to reset this from make commands.