firewall: live view: allow column modifications and combine hostname columns
Since the normal bootgrid implementation expects an action bar to
be able to provide a reset, the shortest path here is to expose
the _setPersistence toggle to allow manual resets from pages
where an action bar is not present.

While here, modify the lookup formatter such that it includes
ports as well, and make sure we unset the src/dst columns
if we lookup hostnames, and vice versa if we don't.