Interfaces: Neighbors: Automatic Discovery - by default, only use non-promiscuous mode, final point for https://github.com/opnsense/core/pull/9354
Added to the roadmap as well, as it seemed to be missing.

This new service is a prerequisite for the new (and faster) mac type aliases (https://github.com/opnsense/core/commit/36dc68554853a0847bf4f71c03b960c49961c247) and is a requirement for IPv6 captive portal in a future release.